pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tribalmind"
version = "0.1.0"
description = "Federated Developer Knowledge Agent - autonomous background agent that observes terminal activity and shares validated fixes across teams"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "TribalMind Contributors" }]

dependencies = [
    "langgraph>=0.4",
    "httpx>=0.27",
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6",
    "keyring>=25",
    "pygithub>=2",
    "platformdirs>=4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
tribal = "tribalmind.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tribalmind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
